+++ ethcore/src/executive_state.rs
@@ -14,6 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity Ethereum.  If not, see <http://www.gnu.org/licenses/>.
 
+//! Execute transactions and modify State. This is glue code between the `ethcore` and
+//! `state-account` crates and contains everything that requires `Machine` or `Executive` (or types
+//! thereof).
+
 use machine::Machine;
 use vm::EnvInfo;
 use executive::{Executive, TransactOptions};
@@ -25,7 +29,7 @@ use types::{
 use trace::{FlatTrace, VMTrace};
 use state_account::{
 	backend::{self, Backend},
-	state::{ApplyResult, ApplyOutcome, State},
+	state::State,
 };
 use ethereum_types::H256;
 use factories::Factories;
@@ -33,6 +37,8 @@ use bytes::Bytes;
 use keccak_hasher::KeccakHasher;
 use kvdb::DBValue;
 use hash_db::AsHashDB;
+
+use error::Error;
 
 // TODO: is there a better place for this?
 /// Return type of proof validity check.
@@ -45,6 +51,21 @@ pub enum ProvedExecution {
 	/// The transaction successfully completed with the given proof.
 	Complete(Box<Executed>),
 }
+
+/// Used to return information about an `State::apply` operation.
+pub struct ApplyOutcome<T, V> {
+	/// The receipt for the applied transaction.
+	pub receipt: Receipt,
+	/// The output of the applied transaction.
+	pub output: Bytes,
+	/// The trace for the applied transaction, empty if tracing was not produced.
+	pub trace: Vec<T>,
+	/// The VM trace for the applied transaction, None if tracing was not produced.
+	pub vm_trace: Option<V>
+}
+
+/// Result type for the execution ("application") of a transaction.
+pub type ApplyResult<T, V> = Result<ApplyOutcome<T, V>, Error>;
 
 /// Check the given proof of execution.
 /// `Err(ExecutionError::Internal)` indicates failure, everything else indicates
@@ -73,7 +94,7 @@ pub fn check_proof(
 	};
 
 	let options = TransactOptions::with_no_tracing().save_output_from_contract();
-	match state.execute(env_info, machine, transaction, options, true) {
+	match execute(&mut state, env_info, machine, transaction, options, true) {
 		Ok(executed) => ProvedExecution::Complete(Box::new(executed)),
 		Err(ExecutionError::Internal(_)) => ProvedExecution::BadProof,
 		Err(e) => ProvedExecution::Failed(e),
@@ -81,7 +102,7 @@ pub fn check_proof(
 }
 
 /// Prove a `virtual` transaction on the given state.
-/// Returns `None` when the transacion could not be proved,
+/// Returns `None` when the transaction could not be proved,
 /// and a proof otherwise.
 pub fn prove_transaction_virtual<H: AsHashDB<KeccakHasher, DBValue> + Send + Sync>(
 	db: H,
@@ -107,7 +128,7 @@ pub fn prove_transaction_virtual<H: AsHashDB<KeccakHasher, DBValue> + Send + Syn
 	};
 
 	let options = TransactOptions::with_no_tracing().dont_check_nonce().save_output_from_contract();
-	match state.execute(env_info, machine, transaction, options, true) {
+	match execute(&mut state, env_info, machine, transaction, options, true) {
 		Err(ExecutionError::Internal(_)) => None,
 		Err(e) => {
 			trace!(target: "state", "Proved call failed: {}", e);
@@ -142,28 +163,18 @@ pub trait ExecutiveStateWithMachineZomgBetterName {
 		where
 			T: trace::Tracer,
 			V: trace::VMTracer;
-	// Execute a given transaction without committing changes.
-	//
-	// `virt` signals that we are executing outside of a block set and restrictions like
-	// gas limits and gas costs should be lifted.
-	fn execute<T, V>(
-		&mut self,
-		env_info: &EnvInfo,
-		machine: &Machine,
-		t: &SignedTransaction,
-		options: TransactOptions<T, V>,
-		virt: bool
-	) -> Result<Executed<T::Output, V::Output>, ExecutionError>
-		where
-			T: trace::Tracer,
-			V: trace::VMTracer;
-
 }
 
 impl<B: Backend> ExecutiveStateWithMachineZomgBetterName for State<B> {
 	/// Execute a given transaction, producing a receipt and an optional trace.
 	/// This will change the state accordingly.
-	fn apply(&mut self, env_info: &EnvInfo, machine: &Machine, t: &SignedTransaction, tracing: bool) -> ApplyResult<FlatTrace, VMTrace> {
+	fn apply(
+		&mut self,
+		env_info: &EnvInfo,
+		machine: &Machine,
+		t: &SignedTransaction,
+		tracing: bool
+	) -> ApplyResult<FlatTrace, VMTrace> {
 		if tracing {
 			let options = TransactOptions::with_tracing();
 			self.apply_with_tracing(env_info, machine, t, options.tracer, options.vm_tracer)
@@ -182,13 +193,13 @@ impl<B: Backend> ExecutiveStateWithMachineZomgBetterName for State<B> {
 		t: &SignedTransaction,
 		tracer: T,
 		vm_tracer: V,
-	) -> ApplyResult<T::Output, V::Output> where
-		T: trace::Tracer,
-		V: trace::VMTracer,
+	) -> ApplyResult<T::Output, V::Output>
+		where
+			T: trace::Tracer,
+			V: trace::VMTracer,
 	{
 		let options = TransactOptions::new(tracer, vm_tracer);
-		// TODO: fix error handling
-		let e = self.execute(env_info, machine, t, options, false).expect("TODO FIXME");
+		let e = execute(self, env_info, machine, t, options, false)?;
 		let params = machine.params();
 
 		let eip658 = env_info.number >= params.eip658_transition;
@@ -218,20 +229,30 @@ impl<B: Backend> ExecutiveStateWithMachineZomgBetterName for State<B> {
 			vm_trace: e.vm_trace,
 		})
 	}
+}
 
-	// Execute a given transaction without committing changes.
-	//
-	// `virt` signals that we are executing outside of a block set and restrictions like
-	// gas limits and gas costs should be lifted.
-	fn execute<T, V>(&mut self, env_info: &EnvInfo, machine: &Machine, t: &SignedTransaction, options: TransactOptions<T, V>, virt: bool)
-	                 -> Result<Executed<T::Output, V::Output>, ExecutionError> where T: trace::Tracer, V: trace::VMTracer,
-	{
-		let schedule = machine.schedule(env_info.number);
-		let mut e = Executive::new(self, env_info, machine, &schedule);
+// Execute a given transaction without committing changes.
+//
+// `virt` signals that we are executing outside of a block set and restrictions like
+// gas limits and gas costs should be lifted.
+fn execute<B, T, V>(
+	state: &mut State<B>,
+	env_info: &EnvInfo,
+	machine: &Machine,
+	t: &SignedTransaction,
+	options: TransactOptions<T, V>,
+	virt: bool
+) -> Result<Executed<T::Output, V::Output>, ExecutionError>
+	where
+		B: Backend,
+		T: trace::Tracer,
+		V: trace::VMTracer,
+{
+	let schedule = machine.schedule(env_info.number);
+	let mut e = Executive::new(state, env_info, machine, &schedule);
 
-		match virt {
-			true => e.transact_virtual(t, options),
-			false => e.transact(t, options),
-		}
+	match virt {
+		true => e.transact_virtual(t, options),
+		false => e.transact(t, options),
 	}
 }
